# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API's MrSoft News",
        "description": "API's for MrSoft News",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://mrsoft-news.test",
            "description": "Servidor de la API (dinámico: usa APP_URL del entorno actual)"
        }
    ],
    "paths": {
...
        "/api/news": {
            "get": {
                "tags": [
                    "News"
                ],
                "summary": "Get all news",
                "description": "Display a listing of the resource.",
                "operationId": "9bfe744cf19dcb58e9ab9677c29ee6c7",
                "parameters": [
                    {
                        "name": "product",
                        "in": "query",
                        "description": "Product name",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "Gesrest",
                                "360sys",
                                "HotelHUB",
                                "Comprobante-e"
                            ]
                        }
                    },
                    {
                        "name": "category",
                        "in": "query",
                        "description": "Category name",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "Noticias",
                                "Eventos",
                                "Promociones"
                            ]
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Limit of news",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NewsRelatedResourceCollection"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Unauthenticated"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "News"
                ],
                "summary": "Create a new news entry",
                "operationId": "1c0fdf6cc53f1508a87121ef9111be5b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description",
                                    "date",
                                    "category_id"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "News title"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "News description"
                                    },
                                    "date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-01-01"
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": "Full news content"
                                    },
                                    "image": {
                                        "type": "string",
                                        "example": "image.jpg"
                                    },
                                    "typeMedia": {
                                        "type": "string",
                                        "example": "video"
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "active": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NewsResource"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/news/{id}": {
            "get": {
                "tags": [
                    "News"
                ],
                "summary": "Get news by id",
                "description": "Display the specified resource.",
                "operationId": "235a202c862ec6888f4129413c1f2bcd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "News id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NewsResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Unauthenticated"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "News"
                ],
                "summary": "Update a news entry",
                "operationId": "11a73338ef61a1ffd2ab3771639f638e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "News id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "News title"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "News description"
                                    },
                                    "date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-01-01"
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": "Full news content"
                                    },
                                    "image": {
                                        "type": "string",
                                        "example": "image.jpg"
                                    },
                                    "typeMedia": {
                                        "type": "string",
                                        "example": "video"
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "active": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NewsResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "News not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "News"
                ],
                "summary": "Delete a news entry",
                "operationId": "cd3235328b3ea2080e4d0a47986d4714",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "News id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "News deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "News not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ValidationError": {
                "properties": {
                    "error": {
                        "type": "string",
                        "example": "The pagination must be an integer."
                    }
                },
                "type": "object"
            },
            "Unauthenticated": {
                "properties": {
                    "error": {
                        "type": "string",
                        "example": "Unauthenticated."
                    }
                },
                "type": "object"
            },
...
            "NewsRelatedResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": "1"
                    },
                    "title": {
                        "type": "string",
                        "example": "News title"
                    },
                    "description": {
                        "type": "string",
                        "example": "News description"
                    },
                    "date": {
                        "type": "string",
                        "example": "2021-09-01"
                    },
                    "category": {
                        "type": "string",
                        "example": "Category name"
                    },
                    "image": {
                        "type": "string",
                        "example": "http://localhost/storage/image.jpg"
                    }
                },
                "type": "object"
            },
            "NewsRelatedResourceCollection": {
                "type": "array",
                "items": {
                    "$ref": "#/components/schemas/NewsRelatedResource"
                }
            },
            "NewsResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": "1"
                    },
                    "title": {
                        "type": "string",
                        "example": "News title"
                    },
                    "description": {
                        "type": "string",
                        "example": "News description"
                    },
                    "date": {
                        "type": "string",
                        "example": "2021-09-01"
                    },
                    "image": {
                        "type": "string",
                        "example": "http://localhost/storage/image.jpg"
                    },
                    "images": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "example": "http://localhost/storage/image.jpg"
                        }
                    },
                    "content": {
                        "type": "string",
                        "example": "News content"
                    },
                    "typeMedia": {
                        "type": "string",
                        "example": "video"
                    },
                    "category": {
                        "type": "string",
                        "example": "Category name"
                    },
                    "category_id": {
                        "type": "integer",
                        "example": "1"
                    },
                    "newsRelated": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/NewsRelatedResource"
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}